Added option to shuffle data for Neural decider
Also option for a control signal architecture so the network is not also
learning whether an option might be valid in a given position or not.
Without control signal, we have 0.0 on an output neuron not
distinguishing between "We have never taken this action" and "This
action has a value of zero".