Two small enhancements for Parameter Search
i) DominionGamesBenchmarkCombination will defaut to true, and will have between 1 and 3 BM deciders in each scoring game. This is to prevent large scores being given to strategies that rely on co-ordination to speed the game to a rapid end.
ii) When using GP to select points, integer parameters will be set at an integer value.